first query switched to react-query
its the query that fills the SNX price graph at the very top
