+++ constants/queryKeys.ts
@@ -2,4 +2,7 @@ export default {
 	Staking: {
 		Liquidations: 'liquidations',
 	},
+	Network: {
+		SnxPriceChart: (chartPeriod: string) => ['snxPriceChart', chartPeriod],
+	},
 };
